requirements.txt
streamlit
numpy
Pillow
onnxruntime
transformers
tokenizers
safetensors